renamed the file and some of the function
